Decompose TaskPanel into focused sub-components
Extract 5 components from the 1472-line monolithic TaskPanel:
- TaskTitleBar: title bar with drag, status dot, action buttons
- TaskBranchInfoBar: branch/project/editor info bar
- TaskNotesPanel: notes/plan tabs + changed files list
- TaskShellSection: shell toolbar + terminal views
- TaskAITerminal: AI agent terminal + exit badge + restart menu

TaskPanel now serves as a coordinator (~300 lines), composing these
sub-components and managing shared state (dialogs, focus, push status).
Each extracted component owns its own signals and lifecycle.
